Update index.html
XSah Coin is a unique community-driven cryptocurrency inspired by the strength and resilience of a ram (🐏) fighting for freedom in the rugged Kurdish mountains. It embodies the spirit of survival alongside nature, symbolizing freedom, honor, and unwavering resistance.

Built on the fast and secure Solana blockchain as an SPL Token, XSah aims to create a decentralized financial ecosystem that honors Kurdish culture and collective struggle.

Join us on this journey to empower a digital symbol of hope, resistance, and cultural pride. XSah is not just a token; it is a movement rooted in heritage and built for the future.